Multiple fixes, working on SQLModel
