+++ backend/{{cookiecutter.project_slug}}/app/main.py
@@ -135,8 +135,9 @@ class LoggingMiddlewareReq(BaseHTTPMiddleware):
                 pass
             response = await call_next(request)
             response_body = await self.resolve_response(response)
-            self.log_request(request, request_body, context)
-            self.log_response(request, response_body)
+            if settings.ENV in ["local", "test"]:
+                self.log_request(request, request_body, context)
+                self.log_response(request, response_body)
             logger_struct.log_struct(
                 {
                     "user": context.get("USER_ID"),
@@ -149,7 +150,6 @@ class LoggingMiddlewareReq(BaseHTTPMiddleware):
                     "content": response_body.body.decode("utf-8"),
                 }
             )
-
         except Exception as e:
             log.error("Error in logging middleware")
             log.error(e)
